feat: Make `edit` command a separate package
